refactor: move info fetching to a separate hook
